+++ pygeodata/drivers/__init__.py
@@ -1,4 +1,5 @@
 from pygeodata.drivers.geopandas import GeoPandasDriver, GeoPandasParquetDriver
+from pygeodata.drivers.pandas import PandasDriver, PandasFormat
 from pygeodata.drivers.rioxarray import RioXArrayDriver
 
-__all__ = ['GeoPandasDriver', 'GeoPandasParquetDriver', 'RioXArrayDriver']
+__all__ = ['GeoPandasDriver', 'GeoPandasParquetDriver', 'PandasDriver', 'PandasFormat', 'RioXArrayDriver']